change the unittests to use oiiotool
