+++ src/cmake/testing.cmake
@@ -302,12 +302,6 @@ macro (oiio_add_all_tests)
                 IMAGEDIR oiio-images
                 ENVIRONMENT "${_pybind_tests_pythonpath}"
                 )
-            oiio_add_tests (
-                raw-thumbnail
-                FOUNDVAR LIBRAW_FOUND ENABLEVAR ENABLE_LIBRAW
-                IMAGEDIR oiio-images/raw
-                ENVIRONMENT "${_pybind_tests_pythonpath}"
-                )
         else ()
             set (nanobind_python_test_suffix "")
         endif ()
@@ -461,7 +455,7 @@ macro (oiio_add_all_tests)
                     IMAGEDIR oiio-images/psd)
     oiio_add_tests (ptex
                     FOUNDVAR Ptex_FOUND ENABLEVAR ENABLE_PTEX)
-    oiio_add_tests (raw
+    oiio_add_tests (raw raw-thumbnail
                     FOUNDVAR LIBRAW_FOUND ENABLEVAR ENABLE_LIBRAW
                     IMAGEDIR oiio-images/raw)
     oiio_add_tests (rla
